Splitting Configuration to separate module
