+++ parity/setup_log.rs
@@ -42,7 +42,7 @@ pub fn setup_log(init: &Option<String>) -> Arc<RotatingLogger> {
 	}
 
 	let logs = Arc::new(RotatingLogger::new(levels));
-	let loggger = logs.clone();
+	let logger = logs.clone();
 	let format = move |record: &LogRecord| {
 		let timestamp = time::strftime("%Y-%m-%d %H:%M:%S %Z", &time::now()).unwrap();
 		let format = if max_log_level() <= LogLevelFilter::Info {
